more refactoring, default init seed with current time, change setSeed

+++ source/bhtsne/include/bhtsne/tsne.h
@@ -79,7 +79,7 @@ public:
     *  @param[in] seed
     *    Random seed
     */
-    void setRandomSeed(int seed);
+    void setRandomSeed(unsigned long seed);
 
     /**
     *  @brief
